+++ plugins/ci/pkg/ci/workload.go
@@ -10,20 +10,20 @@ import (
 
 const workloadsReportVersion = "0.1.0"
 
-func (ci *CIScan) GetWorkloadReport(resources []models.Resource) (models.ReportInfo, error) {
+func (ci *CIScan) GetWorkloadReport(resources []models.Resource) (*models.ReportInfo, error) {
 	workloadsReport := models.ReportInfo{
 		Report:   "scan-workloads",
 		Filename: "scan-workloads.json",
 	}
 	resourceBytes, err := json.Marshal(map[string]interface{}{"Resources": resources})
 	if err != nil {
-		return workloadsReport, err
+		return nil, err
 	}
 	err = os.WriteFile(filepath.Join(ci.config.Options.TempFolder, workloadsReport.Filename), resourceBytes, 0644)
 	if err != nil {
-		return workloadsReport, err
+		return nil, err
 	}
 
 	workloadsReport.Version = workloadsReportVersion
-	return workloadsReport, nil
+	return &workloadsReport, nil
 }